move path related function to `path.rs`
