+++ src/game/utils.rs
@@ -2,7 +2,7 @@ use gb_roms::{
     controllers::{generate_rom_controller, Generic, GenericState, Partial},
     Header,
 };
-use std::{fs::File, path::PathBuf};
+use std::fs::File;
 
 /// Return an initalised MBCs with it auto game save if possible
 pub(crate) fn mbc_with_save_state(
@@ -15,7 +15,7 @@ pub(crate) fn mbc_with_save_state(
     {
         use rmp_serde::decode::from_read;
 
-        let filename = game_save_path(romname);
+        let filename = crate::path::game_save_path(romname);
         if let Ok(file) = File::open(&filename) {
             log::info!("found auto save file at {}", filename.to_string_lossy());
             if let Err(e) =
@@ -35,28 +35,4 @@ pub(crate) fn mbc_with_save_state(
     }
 
     Ok(mbc)
-}
-
-/// Return the path where the game save file will be located
-pub fn game_save_path(rom_filename: &str) -> PathBuf {
-    let rom_id = game_id(rom_filename);
-    let mut root = game_root_config_path();
-
-    root.push(rom_id);
-    root.set_extension(crate::constant::GAME_SAVE_EXT);
-    root
-}
-
-/// Return the root path of the config folder
-pub fn game_root_config_path() -> PathBuf {
-    let mut path = PathBuf::from("~/.config/");
-    path.push(crate::constant::APP_NAME);
-    path
-}
-
-/// Create a standardize rom name id
-fn game_id(rom_filename: &str) -> PathBuf {
-    let mut rom_path = PathBuf::from(rom_filename);
-    rom_path.set_extension("");
-    rom_path
 }
