+++ src/main.rs
@@ -5,6 +5,7 @@ mod custom_event;
 mod game;
 mod image;
 mod logger;
+mod path;
 #[cfg(any(feature = "time_frame", feature = "debug_fps"))]
 mod time_frame;
 mod ui;
